backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
supabase==2.4.6
python-multipart==0.0.9
httpx==0.27.0
python-jose==3.3.0
passlib==1.7.4
google-generativeai==0.5.4
groq==0.9.0
pandas==2.2.2